refactor:change some styles
