+++ aboutCompany/src/components/aboutPage/AboutUs.tsx
