+++ aboutCompany/src/components/companyDescription/aboutSection.module.scss
