+++ aboutCompany/src/components/header/Header.module.scss
@@ -1,6 +1,6 @@
 .headerWrapper {
   max-width: 1200px;
-  padding: 35px 0px;
+  padding: 5px 0px;
   margin: 0 auto;
 }
