+++ aboutCompany/src/components/main/Main.tsx
@@ -1,16 +1,16 @@
 import React from "react";
 import { FollowUs } from "../follow/FollowUs";
-import { AboutSection } from "../aboutSection/aboutSection";
-import { CollectionSection } from "../collectionsSection/CollectionSection";
-import { ContactUs } from "../contacts/Contacts";
+import { CompanyDescription } from "../companyDescription/aboutSection";
+import { ServicesInfo } from "../servicesList/CollectionSection";
+import { ContactInfo } from "../contactInfo/Contacts";
 
 export const Main: React.FC = () => {
   return (
     <main>
       <FollowUs />
-      <AboutSection />
-      <CollectionSection />
-      <ContactUs />
+      <CompanyDescription />
+      <ServicesInfo />
+      <ContactInfo />
     </main>
   );
 };
