+++ aboutCompany/src/components/router/Router.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import { BrowserRouter, Routes, Route } from "react-router-dom";
-import { About } from "../about/AboutUs";
+import { About } from "../aboutPage/AboutUs";
 import { SignUp } from "../signUp/SignUp";
 import { Collection } from "../collection/Collection";
 import { Header } from "../header/Header";
@@ -17,7 +17,7 @@ export const Router: React.FC = () => {
         <Route path="/signIn" element={<SignIn />} />
         <Route path="/signUp" element={<SignUp />} />
         <Route path="/collection" element={<Collection />} />
-        <Route path="/aboutUs" element={<About />} />
+        <Route path="/about" element={<About />} />
       </Routes>
       <Footer />
     </BrowserRouter>
